break Plotly.react tests out of plot_api_test into plot_api_react_test
